Gauge no longer has a timeout Gauge that didn't change used to disappear from results; this could make a non-zero value appear zero, confusing metrics. gaugeCommon.go split was confusing, merge back into gauge.go
